[ipkg] * move ipkg filescan from Softwaremanager plugin to scanner
